requirements.txt
pytest
google-generativeai
python-dotenv
requests
